requirements.txt
Flask==2.3.3
Pillow==10.0.1
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
Flask-Limiter==3.5.0
Flask-CORS==4.0.0
Flask-RESTX==1.2.0
opencv-python==4.8.1.78
numpy==1.24.3
twilio==8.10.0
